+++ java/core/src/main/java/sleeper/core/tracker/ingest/job/update/IngestJobAddedFilesEvent.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.ingest.core.job.status;
+package sleeper.core.tracker.ingest.job.update;
 
 import sleeper.core.statestore.AllReferencesToAFile;
 
@@ -22,7 +22,7 @@ import java.util.List;
 import java.util.Objects;
 
 /**
- * An event for when an ingest job committed files to the state store. Used in the ingest job status store.
+ * An event for when an ingest job committed files to the state store. Used in the ingest job tracker.
  */
 public class IngestJobAddedFilesEvent {
     private final String jobId;
