+++ java/core/src/main/java/sleeper/core/tracker/ingest/job/update/IngestJobFailedEvent.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.ingest.core.job.status;
+package sleeper.core.tracker.ingest.job.update;
 
 import sleeper.core.record.process.ProcessRunTime;
 
@@ -22,7 +22,7 @@ import java.util.List;
 import java.util.Objects;
 
 /**
- * An event for when an ingest job failed. Used in the ingest job status store.
+ * An event for when an ingest job failed. Used in the ingest job tracker.
  */
 public class IngestJobFailedEvent {
     private final String jobId;
