+++ java/ingest/ingest-core/src/main/java/sleeper/ingest/core/job/status/IngestJobStatusStore.java
@@ -16,6 +16,11 @@
 
 package sleeper.ingest.core.job.status;
 
+import sleeper.core.tracker.ingest.job.update.IngestJobAddedFilesEvent;
+import sleeper.core.tracker.ingest.job.update.IngestJobFailedEvent;
+import sleeper.core.tracker.ingest.job.update.IngestJobFinishedEvent;
+import sleeper.core.tracker.ingest.job.update.IngestJobStartedEvent;
+
 import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
