+++ java/ingest/ingest-runner/src/main/java/sleeper/ingest/runner/impl/commit/AddFilesToStateStore.java
@@ -27,9 +27,9 @@ import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
 import sleeper.core.statestore.commit.StateStoreCommitRequestInS3Uploader;
+import sleeper.core.tracker.ingest.job.update.IngestJobAddedFilesEvent;
 import sleeper.ingest.core.job.commit.IngestAddFilesCommitRequest;
 import sleeper.ingest.core.job.commit.IngestAddFilesCommitRequestSerDe;
-import sleeper.ingest.core.job.status.IngestJobAddedFilesEvent;
 import sleeper.ingest.core.job.status.IngestJobStatusStore;
 
 import java.util.List;
